updatig user model with additional constructor

+++ src/main/java/ea/sof/shared/models/Auth.java
@@ -10,14 +10,14 @@ import lombok.Setter;
 @NoArgsConstructor
 public class Auth {
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    private Integer userId;
+    private Long userId;
     private String email;
     private String password;
